started cleaning up repository layer for wb
